+++ build.gradle
@@ -7,6 +7,7 @@ version 'master'
 repositories {
     mavenCentral()
     jcenter()
+    maven { url 'https://jitpack.io' }
 }
 
 dependencies {
@@ -15,6 +16,7 @@ dependencies {
     compile group: 'ws.schild', name: 'jave-all-deps', version: '2.7.3'
     implementation("com.google.guava:guava:28.2-jre")
     compile group: 'ch.qos.logback', name: 'logback-classic', version: '1.2.3'
+    implementation 'com.github.DiscordBotList:DBL-Java-Library:2.0.1'
 }
 
 compileKotlin {
